+++ core/src/test/java/uk/gov/justice/services/core/it/EventHandlerIT.java
@@ -64,11 +64,10 @@ import uk.gov.justice.services.core.requester.RequesterProducer;
 import uk.gov.justice.services.core.sender.SenderProducer;
 import uk.gov.justice.services.messaging.DefaultJsonObjectEnvelopeConverter;
 import uk.gov.justice.services.messaging.JsonEnvelope;
-import uk.gov.justice.services.messaging.jms.DefaultEnvelopeConverter;
-import uk.gov.justice.services.messaging.jms.DefaultJmsEnvelopeSender;
 import uk.gov.justice.services.messaging.logging.DefaultTraceLogger;
 import uk.gov.justice.services.test.utils.common.envelope.TestEnvelopeRecorder;
 import uk.gov.justice.services.test.utils.common.validator.DummyJsonSchemaValidator;
+import uk.gov.justice.services.test.utils.messaging.jms.DummyJmsEnvelopeSender;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -118,8 +117,7 @@ public class EventHandlerIT {
             EventListenerInterceptorChainProvider.class,
 
             SenderProducer.class,
-            DefaultJmsEnvelopeSender.class,
-            DefaultEnvelopeConverter.class,
+            DummyJmsEnvelopeSender.class,
             DispatcherFactory.class,
             EnvelopePayloadTypeConverter.class,
             JsonEnvelopeRepacker.class,