+++ core/src/test/java/uk/gov/justice/services/core/it/SenderRequesterHandlerIT.java
@@ -59,11 +59,10 @@ import uk.gov.justice.services.core.sender.Sender;
 import uk.gov.justice.services.core.sender.SenderProducer;
 import uk.gov.justice.services.messaging.DefaultJsonObjectEnvelopeConverter;
 import uk.gov.justice.services.messaging.JsonEnvelope;
-import uk.gov.justice.services.messaging.jms.DefaultEnvelopeConverter;
-import uk.gov.justice.services.messaging.jms.DefaultJmsEnvelopeSender;
 import uk.gov.justice.services.messaging.logging.DefaultTraceLogger;
 import uk.gov.justice.services.test.utils.common.envelope.TestEnvelopeRecorder;
 import uk.gov.justice.services.test.utils.common.validator.DummyJsonSchemaValidator;
+import uk.gov.justice.services.test.utils.messaging.jms.DummyJmsEnvelopeSender;
 
 import java.util.UUID;
 
@@ -119,8 +118,12 @@ public class SenderRequesterHandlerIT {
             ServiceComponentObserver.class,
             RequesterProducer.class,
             SenderProducer.class,
-            DefaultJmsEnvelopeSender.class,
-            DefaultEnvelopeConverter.class,
+
+            DummyJmsEnvelopeSender.class,
+            DispatcherFactory.class,
+            EnvelopePayloadTypeConverter.class,
+            JsonEnvelopeRepacker.class,
+
             StringToJsonObjectConverter.class,
             DefaultJsonObjectEnvelopeConverter.class,
             ObjectToJsonValueConverter.class,