+++ messaging/messaging-core/src/test/java/uk/gov/justice/services/messaging/JsonEnvelopeWriterTest.java
@@ -8,7 +8,6 @@ import javax.json.JsonObject;
 
 import org.junit.Test;
 import org.junit.runner.RunWith;
-import org.mockito.InjectMocks;
 import org.mockito.runners.MockitoJUnitRunner;