+++ messaging/messaging-core/src/test/java/uk/gov/justice/services/messaging/logging/DefaultTraceLoggerTest.java
@@ -1,29 +1,16 @@
 package uk.gov.justice.services.messaging.logging;
 
-import static org.junit.Assert.*;
-
-import org.junit.runner.RunWith;
-import org.mockito.internal.matchers.Any;
-import org.mockito.runners.MockitoJUnitRunner;
-
-import static org.hamcrest.CoreMatchers.is;
-import static org.hamcrest.CoreMatchers.sameInstance;
-import static org.junit.Assert.assertThat;
+import static org.mockito.Mockito.any;
+import static org.mockito.Mockito.mock;
+import static org.mockito.Mockito.never;
+import static org.mockito.Mockito.verify;
+import static org.mockito.Mockito.verifyZeroInteractions;
 import static org.mockito.Mockito.when;
 
-import org.junit.Test;
-import org.junit.runner.RunWith;
-import org.mockito.InjectMocks;
-import org.mockito.Mock;
-import org.mockito.runners.MockitoJUnitRunner;
-import org.slf4j.Logger;
-
-import static org.hamcrest.CoreMatchers.equalTo;
-import static org.hamcrest.CoreMatchers.is;
-import static org.junit.Assert.assertThat;
-import static org.mockito.Mockito.*;
-
 import java.util.function.Supplier;
+
+import org.junit.Test;
+import org.slf4j.Logger;
 
 @SuppressWarnings("unchecked")
 public class DefaultTraceLoggerTest {