+++ messaging/messaging-jms/src/main/java/uk/gov/justice/services/messaging/jms/EnvelopeSender.java
@@ -1,0 +1,7 @@
+package uk.gov.justice.services.messaging.jms;
+
+import uk.gov.justice.services.messaging.JsonEnvelope;
+
+public interface EnvelopeSender {
+    void send(final JsonEnvelope command, final String destinationName);
+}